Made some changes

+++ Views/Account.js
@@ -1,16 +1,16 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useState, useCallback } from "react";
 import {
   View,
   Text,
-  Button,
   Image,
   StyleSheet,
   TouchableOpacity,
+  ScrollView,
 } from "react-native";
 import { doc, getDoc, deleteDoc } from "firebase/firestore";
 import { getAuth, deleteUser } from "firebase/auth";
 import { getStorage, ref, getDownloadURL } from "firebase/storage";
-import { useNavigation, useRoute } from "@react-navigation/native";
+import { useNavigation, useFocusEffect } from "@react-navigation/native";
 import { FIRESTORE_DB, FIREBASE_AUTH } from "../FirebaseConfig";
 
 function Account() {
@@ -18,10 +18,12 @@ function Account() {
   const [currentUser, setCurrentUser] = useState(null);
   const [profileImage, setProfileImage] = useState(null);
 
-  useEffect(() => {
-    fetchUserData();
-    fetchProfileImage();
-  }, []);
+  useFocusEffect(
+    useCallback(() => {
+      fetchUserData();
+      fetchProfileImage();
+    }, [])
+  );
 
   const fetchUserData = async () => {
     try {
@@ -62,14 +64,15 @@ function Account() {
   const deleteUserAccount = async () => {
     try {
       const userAuth = getAuth().currentUser;
+      if (userAuth) {
+        await deleteDoc(doc(FIRESTORE_DB, "users", userAuth.uid));
+        await deleteUser(userAuth);
 
-      // Delete the user document from Firestore
-      await deleteDoc(doc(FIRESTORE_DB, "users", userAuth.uid));
-
-      // Delete the user from Firebase Authentication
-      await deleteUser(userAuth);
-
-      navigation.navigate("Login");
+        console.log("User account deleted successfully");
+        navigation.navigate("Login");
+      } else {
+        console.log("No authenticated user found");
+      }
     } catch (error) {
       console.log("Error deleting user account:", error);
     }
@@ -85,101 +88,230 @@ function Account() {
   };
 
   return (
-    <View style={styles.container}>
-      {profileImage && (
-        <Image source={{ uri: profileImage }} style={styles.profileImage} />
-      )}
-      <Text style={styles.heading}>Account Information:</Text>
-      {currentUser ? (
-        <>
-          <Text style={styles.text}>
-            Username: {currentUser.username || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            Email: {FIREBASE_AUTH.currentUser?.email || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            Address: {currentUser.address || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            Age: {currentUser.age || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            First Name: {currentUser.firstName || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            Last Name: {currentUser.lastName || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            Occupation: {currentUser.occupation || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            Room: {currentUser.room || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            Others: {currentUser.others || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            Family: {currentUser.family || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            Salary: {currentUser.salary || "Not filled"}
-          </Text>
-          <Text style={styles.text}>
-            Time: {currentUser.time || "Not filled"}
-          </Text>
-        </>
-      ) : (
-        <Text>No user data available</Text>
-      )}
+    <ScrollView style={styles.container}>
+      <View style={styles.header}>
+        {profileImage && (
+          <Image source={{ uri: profileImage }} style={styles.profileImage} />
+        )}
+        {currentUser && (
+          <View style={styles.userInfo}>
+            <Text style={styles.userName}>
+              {currentUser.firstName} {currentUser.lastName}
+            </Text>
+            <Text style={styles.userOccupation}>
+              {currentUser.occupation || "Occupation not filled"}
+            </Text>
+          </View>
+        )}
+      </View>
 
-      <Button title="Delete User" onPress={deleteUserAccount} />
-      <Button title="Sign Out" onPress={signOutUser} />
+      <View style={styles.actions}>
+        <TouchableOpacity style={styles.actionButton}>
+          <Text style={styles.actionText}>Email</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.actionButton}
+          onPress={() => navigation.navigate("Messages")}
+        >
+          <Text style={styles.actionText}>Chats</Text>
+        </TouchableOpacity>
+        <TouchableOpacity style={styles.actionButton}>
+          <Text style={styles.actionText}>Favorite</Text>
+        </TouchableOpacity>
+      </View>
 
-      <TouchableOpacity
-        style={styles.nextButton}
-        onPress={() => navigation.navigate("Chat", { userId: userId })}
-      >
-        <Text style={styles.nextButtonText}>Next</Text>
-      </TouchableOpacity>
+      <View style={styles.details}>
+        <Text style={styles.detailTitle}>Email</Text>
+        <Text style={styles.detailText}>
+          {FIREBASE_AUTH.currentUser?.email || "Not filled"}
+        </Text>
 
-      <TouchableOpacity
-        style={styles.nextButton}
-        onPress={() => navigation.navigate("Customer Profiles")}
-      >
-        <Text>Customer</Text>
-      </TouchableOpacity>
+        <Text style={styles.detailTitle}>Address</Text>
+        <Text style={styles.detailText}>
+          {currentUser?.address || "Not filled"}
+        </Text>
 
-      <TouchableOpacity
-        style={styles.nextButton}
-        onPress={() => navigation.navigate("Messages")}
-      >
-        <Text>Chats</Text>
-      </TouchableOpacity>
-    </View>
+        <Text style={styles.detailTitle}>Age</Text>
+        <Text style={styles.detailText}>
+          {currentUser?.age || "Not filled"}
+        </Text>
+
+        {currentUser?.occupation === "cleaner" ? (
+          <>
+            <Text style={styles.detailTitle}>Years of Experience</Text>
+            <Text style={styles.detailText}>
+              {currentUser?.experience || "Not filled"}
+            </Text>
+
+            <Text style={styles.detailTitle}>Expected Salary</Text>
+            <Text style={styles.detailText}>
+              {currentUser?.salary || "Not filled"}
+            </Text>
+
+            <Text style={styles.detailTitle}>Hours Wanted to work</Text>
+            <Text style={styles.detailText}>
+              {currentUser?.work || "Not filled"}
+            </Text>
+
+            <Text style={styles.detailTitle}>Time Slot</Text>
+            <Text style={styles.detailText}>
+              {currentUser?.time || "Not filled"}
+            </Text>
+
+            <Text style={styles.detailTitle}>Bio</Text>
+            <Text style={styles.detailText}>
+              {currentUser?.info || "Not filled"}
+            </Text>
+          </>
+        ) : (
+          <>
+            <Text style={styles.detailTitle}>Family</Text>
+            <Text style={styles.detailText}>
+              {currentUser?.family || "Not filled"}
+            </Text>
+
+            <Text style={styles.detailTitle}>Room</Text>
+            <Text style={styles.detailText}>
+              {currentUser?.room || "Not filled"}
+            </Text>
+
+            <Text style={styles.detailTitle}>Others</Text>
+            <Text style={styles.detailText}>
+              {currentUser?.others || "Not filled"}
+            </Text>
+
+            <Text style={styles.detailTitle}>Salary</Text>
+            <Text style={styles.detailText}>
+              {currentUser?.salary || "Not filled"}
+            </Text>
+          </>
+        )}
+      </View>
+
+      <View style={styles.profileButtons}>
+        <TouchableOpacity
+          style={styles.profileButton}
+          onPress={() => navigation.navigate("Cleaner Profiles")}
+        >
+          <Text style={styles.profileButtonText}>View Cleaners</Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={styles.profileButton}
+          onPress={() => navigation.navigate("Customer Profiles")}
+        >
+          <Text style={styles.profileButtonText}>View Customers</Text>
+        </TouchableOpacity>
+      </View>
+
+      <View style={styles.footer}>
+        <TouchableOpacity
+          style={styles.deleteButton}
+          onPress={deleteUserAccount}
+        >
+          <Text style={styles.deleteButtonText}>Delete User</Text>
+        </TouchableOpacity>
+        <TouchableOpacity style={styles.signOutButton} onPress={signOutUser}>
+          <Text style={styles.signOutButtonText}>Sign Out</Text>
+        </TouchableOpacity>
+      </View>
+    </ScrollView>
   );
 }
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
+    backgroundColor: "#fff",
+  },
+  header: {
+    backgroundColor: "#4A90E2",
+    paddingVertical: 30,
     alignItems: "center",
-    justifyContent: "center",
-  },
-  heading: {
-    fontSize: 18,
-    fontWeight: "bold",
-    marginVertical: 10,
-  },
-  text: {
-    fontSize: 16,
-    marginBottom: 5,
   },
   profileImage: {
-    width: 200,
-    height: 200,
-    borderRadius: 100,
-    marginBottom: 20,
+    width: 100,
+    height: 100,
+    borderRadius: 50,
+  },
+  userInfo: {
+    marginTop: 10,
+    alignItems: "center",
+  },
+  userName: {
+    fontSize: 20,
+    fontWeight: "bold",
+    color: "#fff",
+  },
+  userOccupation: {
+    fontSize: 16,
+    color: "#fff",
+  },
+  actions: {
+    flexDirection: "row",
+    justifyContent: "space-around",
+    marginVertical: 20,
+  },
+  actionButton: {
+    alignItems: "center",
+    backgroundColor: "#4A90E2",
+    padding: 10,
+    borderRadius: 10,
+  },
+  actionText: {
+    color: "#fff",
+  },
+  details: {
+    paddingHorizontal: 20,
+  },
+  detailTitle: {
+    fontSize: 16,
+    fontWeight: "bold",
+    marginTop: 20,
+  },
+  detailText: {
+    fontSize: 16,
+    color: "#333",
+    marginTop: 5,
+  },
+  profileButtons: {
+    flexDirection: "row",
+    justifyContent: "space-around",
+    marginVertical: 20,
+  },
+  profileButton: {
+    alignItems: "center",
+    backgroundColor: "#4A90E2",
+    padding: 10,
+    borderRadius: 5,
+    flex: 1,
+    marginHorizontal: 10,
+  },
+  profileButtonText: {
+    color: "#fff",
+  },
+  footer: {
+    marginVertical: 20,
+    alignItems: "center",
+  },
+  deleteButton: {
+    marginVertical: 10,
+    padding: 10,
+    backgroundColor: "red",
+    borderRadius: 5,
+  },
+  deleteButtonText: {
+    color: "#fff",
+    fontWeight: "bold",
+  },
+  signOutButton: {
+    marginVertical: 10,
+    padding: 10,
+    backgroundColor: "#4A90E2",
+    borderRadius: 5,
+  },
+  signOutButtonText: {
+    color: "#fff",
+    fontWeight: "bold",
   },
 });
